refactor(path): Improved path organization in all files
