+++ server/src/domain/datasources/pizza.datasource.ts
@@ -1,6 +1,5 @@
-import { PizzaIdDto } from "../dtos/pizza/pizza-id.dto";
-import { PizzaDto } from "../dtos/pizza/pizza.dto";
-import { PizzaEntity } from "../entities/pizza.entity";
+import { PizzaDto, PizzaIdDto, PizzaEntity } from "../index";
+
 export abstract class PizzaDatasource {
     abstract create(pizzaDto: PizzaDto): Promise<PizzaEntity>;
     abstract getAll(): Promise<PizzaEntity[]>;
