+++ server/src/domain/dtos/index.ts
@@ -1,0 +1,1 @@
+export * from "./pizza/index";
