+++ server/src/domain/dtos/pizza/index.ts
@@ -1,0 +1,2 @@
+export * from "./pizza-id.dto";
+export * from "./pizza.dto";
