+++ server/src/domain/entities/index.ts
@@ -1,0 +1,1 @@
+export * from "./pizza.entity";
