+++ server/src/domain/index.ts
@@ -1,5 +1,6 @@
-export * from "./entities/pizza.entity";
-export * from "./dtos/pizza/pizza.dto";
-export * from "./datasources/pizza.datasource";
-export * from "./repositories/pizza.repository";
 export * from "./errors/custom.error";
+export * from "./entities/index";
+export * from "./dtos/index";
+export * from "./datasources/index";
+export * from "./repositories/index";
+export * from "./use-cases/index";
