+++ server/src/domain/repositories/pizza.repository.ts
@@ -1,6 +1,5 @@
-import { PizzaIdDto } from "../dtos/pizza/pizza-id.dto";
-import { PizzaDto } from "../dtos/pizza/pizza.dto";
-import { PizzaEntity } from "../entities/pizza.entity";
+import { PizzaIdDto, PizzaDto, PizzaEntity } from "../index";
+
 export abstract class PizzaRepository {
     abstract create(PizzaDto: PizzaDto): Promise<PizzaEntity>;
     abstract getAll(): Promise<PizzaEntity[]>;
