+++ server/src/infrastructure/index.ts
@@ -1,3 +1,3 @@
-export * from "./datasources/pizza.datasource.impl";
-export * from "./respositories/pizza.repository.impl";
-export * from "./mappers/pizza.mapper";
+export * from "./datasources/index";
+export * from "./respositories/index";
+export * from "./mappers/index";
