+++ server/src/infrastructure/mappers/index.ts
@@ -1,0 +1,1 @@
+export * from "./pizza.mapper";
