+++ server/src/infrastructure/respositories/index.ts
@@ -1,0 +1,1 @@
+export * from "./pizza.repository.impl";
