+++ server/src/presentation/products/v1/beverage/index.ts
@@ -1,0 +1,1 @@
+export * from "./beverage.routes";
