+++ server/src/presentation/products/v1/dessert/index.ts
@@ -1,0 +1,1 @@
+export * from "./dessert.routes";
