+++ server/src/presentation/products/v1/empanada/index.ts
@@ -1,0 +1,1 @@
+export * from "./empanada.routes";
