+++ server/src/presentation/products/v1/index.ts
@@ -1,4 +1,4 @@
-export * from "./pizza/pizza.routes";
-export * from "./beverage/beverage.routes";
-export * from "./dessert/dessert.routes";
-export * from "./empanada/empanada.routes";
+export * from "./pizza/index";
+export * from "./beverage/index";
+export * from "./dessert/index";
+export * from "./empanada/index";
