+++ server/src/presentation/products/v1/pizza/index.ts
@@ -1,0 +1,2 @@
+export * from "./pizza.controller";
+export * from "./pizza.routes";
